27. Remove Elements
'''Given an integer array nums and an integer val, remove all occurrences of val in nums in-place. The order of the elements may be changed. Then return the number of elements in nums which are not equal to val.

Consider the number of elements in nums which are not equal to val be k, to get accepted, you need to do the following things:

Change the array nums such that the first k elements of nums contain the elements which are not equal to val. The remaining elements of nums are not important as well as the size of nums.
Return k.'''

#Code
def removeElement(self, nums, val):
    nums[:] = [e for e in nums if e != val]
    return len(nums)
    
"
In the line nums[:] is using slice assignment. This means that the contents of the nums list 
are replaced with the new list comprehension [e for e in nums if e != val]. 
The slice assignment modifies the existing list in place, effectively changing its content to the filtered values.
"